feat!(refactor): Breaking change to separate client and server code
This is to allow the server to exist as it's own entity and to potentially be versioned independently from the client, the same goes for the client, or treecli. Additionally the proto files have been moved into their own directory as well as they will be shared between both the client and server. Finally dummy data has been added, although this may be better suited to residing in the server directory as it will mostly be used for integration tests